+++ lib/document.dart
@@ -1,12 +1,12 @@
 library document;
 
-export 'package:json_api_common/src/document/document.dart';
 export 'package:json_api_common/src/document/error_object.dart';
 export 'package:json_api_common/src/document/identifier.dart';
 export 'package:json_api_common/src/document/identity.dart';
+export 'package:json_api_common/src/document/inbound_document.dart';
 export 'package:json_api_common/src/document/link.dart';
 export 'package:json_api_common/src/document/new_resource.dart';
-export 'package:json_api_common/src/document/relationship/incomplete.dart';
+export 'package:json_api_common/src/document/outbound_document.dart';
 export 'package:json_api_common/src/document/relationship/many.dart';
 export 'package:json_api_common/src/document/relationship/one.dart';
 export 'package:json_api_common/src/document/relationship/relationship.dart';