+++ lib/http.dart
@@ -1,7 +1,6 @@
-/// This is a thin HTTP layer abstraction used by the client and the server.
+/// This is a thin HTTP layer abstraction used by the client
 library http;
 
-export 'package:json_api_common/src/http/content_type.dart';
 export 'package:json_api_common/src/http/headers.dart';
 export 'package:json_api_common/src/http/http_handler.dart';
 export 'package:json_api_common/src/http/http_request.dart';