+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.5] - 2020-11-02
+### Changed
+- Document got split into the Inbound and Outbound documents
+
+### Removed
+- Target objects from the Routing
+
 ## [0.0.4] - 2020-10-19
 ### Added
 - Document library
@@ -25,5 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial version.
 
-[Unreleased]: https://github.com/f3ath/jessie/json-api-common/0.0.4...HEAD
+[Unreleased]: https://github.com/f3ath/jessie/json-api-common/0.0.5...HEAD
+[0.0.5]: https://github.com/f3ath/jessie/json-api-common/0.0.4...0.0.5
 [0.0.4]: https://github.com/f3ath/jessie/json-api-common/0.0.3...0.0.4